feat: Added the correct annotations to the model classes and changed the invested attribute to be a double instead of an int

+++ src/main/java/com/insightfindr/model/Stock.java
@@ -27,7 +27,7 @@ public class Stock {
     private String message;
 
     @Column(name = "amount_invested")
-    private int invested;
+    private double invested;
 
     @Column(name = "date_invested")
     private String date_invested;
